+++ source/sheet.cpp
@@ -10,7 +10,7 @@ Sheet::Sheet(std::string t_name, uint8_t t_strength, uint8_t t_dexterity,
     m_name( std::move(t_name) ), m_strength(t_strength),
     m_dexterity(t_dexterity),  m_constitution(t_constitution),
     m_inteligence(t_inteligence), m_wisdom(t_wisdom), m_charisma(t_charisma),
-    m_level(t_level), m_maxLife(t_maxLife), m_currentLife(t_maxLife)
+    m_level(t_level),m_maxLife(t_maxLife), m_currentLife(t_maxLife)
 {
 }
 
@@ -18,23 +18,23 @@ Sheet::~Sheet() = default;
 
 const std::string &Sheet::get_name() const { return m_name; }
 
-uint8_t Sheet::get_level() const { return m_level; }
+int Sheet::get_level() const { return m_level; }
 
-uint8_t Sheet::get_strength() const { return m_strength; }
+int Sheet::get_strength() const { return m_strength; }
 
-uint8_t Sheet::get_dexterity() const { return m_dexterity; }
+int Sheet::get_dexterity() const { return m_dexterity; }
 
-uint8_t Sheet::get_constitution() const { return m_constitution; }
+int Sheet::get_constitution() const { return m_constitution; }
 
-uint8_t Sheet::get_inteligence() const { return m_inteligence; }
+int Sheet::get_inteligence() const { return m_inteligence; }
 
-uint8_t Sheet::get_wisdom() const { return m_wisdom; }
+int Sheet::get_wisdom() const { return m_wisdom; }
 
-uint8_t Sheet::get_charisma() const { return m_charisma; }
+int Sheet::get_charisma() const { return m_charisma; }
 
-unsigned short Sheet::get_maxLife() const { return m_maxLife; }
+int Sheet::get_maxLife() const { return m_maxLife; }
 
-unsigned short Sheet::get_currentLife() const { return m_currentLife; }
+int Sheet::get_currentLife() const { return m_currentLife; }
 
 Weapon& Sheet::get_naturalWeapon() const { return *m_naturalWeapon; }
